Actually use the SierraMergeCandidates trait
