+++ catalogue_pipeline/transformer/src/test/scala/uk/ac/wellcome/platform/transformer/transformers/sierra/SierraMergeCandidatesTest.scala
@@ -100,23 +100,3 @@ class SierraMergeCandidatesTest extends FunSpec with Matchers with SierraData {
     transformer.getMergeCandidates(sierraData) shouldBe Nil
   }
 }
-
-trait SierraMergeCandidates extends MarcUtils{
-
-  def getMergeCandidates(sierraBibData: SierraBibData): List[MergeCandidate] = for {
-    subFields <- getMatchingSubfields(sierraBibData, "776", "w")
-    subField <- subFields
-    mergeCandidate <- extractMergeCandidate(subField.content).toList
-  } yield mergeCandidate
-
-  private def extractMergeCandidate(content: String): Option[MergeCandidate] = {
-    // This regex matches any string starting with (UkLW), followed by any number of spaces
-    // and groups everything after that, which is the sierra bib number we're interested in
-    val prefixRegex = """\(UkLW\)[\s]*(.+)""".r.anchored
-
-    content match {
-      case prefixRegex(bibNumber) => Some(MergeCandidate(SourceIdentifier(IdentifierType("sierra-system-number"), "Work", bibNumber)))
-      case _ => None
-    }
-  }
-}
